Fix pacman and ghosts unable to re-enter maze after tunnel teleportation
This only occured when an actor teleported to the other side of the
tunnel whilst the outer tunnel door was locked. The actor could not
enter the maze until the doors were opened. For ghosts it was not
a big problem since the doors were opened by pacman as he ate keys.
But for the player the game became unplayable since they could no
longer get to the keys